CloudRun/pyproject.toml
[tool.poetry]
name = "cloudrun"
version = "0.1.0"
description = ""
authors = ["FujiiNoritsugu <quercus491@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.45.0"
google-cloud-aiplatform = "^1.38.1"
mecab-python3 = "^1.0.6"
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
vertexai = "^0.0.1"
scipy = "^1.11.4"
wikipedia = "^1.4.0"
unidic-lite = "^1.0.8"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
flake8 = "^7.2.0"
pip-audit = "^2.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
